+++ BrewTests/BrewAPIClientCatalogueURLSessionIntegrationTests.swift
@@ -117,115 +117,4 @@ struct BrewAPICatalogueIntegrationTests {
             #expect(data.decodeFailures.first?.index == 0)
         }
     }
-
-    @Test func `decodes homebrew wire cask bulk shape`() throws {
-        let data = Data(
-            """
-            [
-              {
-                "token": "iterm2",
-                "name": ["iTerm2"],
-                "desc": "Terminal emulator",
-                "homepage": "https://iterm2.com",
-                "version": "3.5.0",
-                "depends_on": { "macos": { ">=": ["12"] } }
-              }
-            ]
-            """.utf8,
-        )
-
-        let decoded = try JSONDecoder().decode(CaskCatalogueJSON.self, from: data)
-
-        #expect(decoded.items.count == 1)
-        #expect(decoded.decodeFailures.isEmpty)
-        #expect(decoded.items.first?.name == "iterm2")
-        #expect(decoded.items.first?.displayName == "iTerm2")
-        #expect(decoded.items.first?.stableVersion == "3.5.0")
-    }
-
-    @Test func `decodes cask with null desc`() throws {
-        let data = Data(
-            """
-            [
-              {
-                "token": "0-ad",
-                "name": ["0 A.D."],
-                "desc": null,
-                "homepage": "https://play0ad.com/",
-                "version": "0.28.0",
-                "depends_on": { "macos": {} }
-              }
-            ]
-            """.utf8,
-        )
-
-        let decoded = try JSONDecoder().decode(CaskCatalogueJSON.self, from: data)
-
-        #expect(decoded.items.count == 1)
-        #expect(decoded.decodeFailures.isEmpty)
-        #expect(decoded.items.first?.description == nil)
-    }
-
-    @Test func `decodes cask with omitted desc`() throws {
-        let data = Data(
-            """
-            [
-              {
-                "token": "0-ad",
-                "name": ["0 A.D."],
-                "homepage": "https://play0ad.com/",
-                "version": "0.28.0",
-                "depends_on": { "macos": {} }
-              }
-            ]
-            """.utf8,
-        )
-
-        let decoded = try JSONDecoder().decode(CaskCatalogueJSON.self, from: data)
-
-        #expect(decoded.items.count == 1)
-        #expect(decoded.decodeFailures.isEmpty)
-        #expect(decoded.items.first?.description == nil)
-    }
-
-    @Test @MainActor func `fetch formula catalogue keeps valid items when one item fails decoding`() async throws {
-        let baseURL = makeStubBaseURL()
-        try StubURLProtocol.register(
-            [
-                .successWithStatus(
-                    data: Data(
-                        """
-                        [
-                          {
-                            "name": "wget",
-                            "desc": "Network downloader",
-                            "homepage": "https://www.gnu.org/software/wget/",
-                            "versions": { "stable": "1.24.5" },
-                            "dependencies": []
-                          },
-                          {
-                            "name": "broken-item"
-                          }
-                        ]
-                        """.utf8,
-                    ),
-                    statusCode: 200,
-                ),
-            ],
-            forHost: #require(baseURL.host),
-        )
-        let session = makeStubbedSession()
-        let client = URLSessionBrewAPIClient(session: session, baseURL: baseURL)
-
-        let response = try await client.fetchFormulaCatalogue(etag: nil)
-        switch response {
-        case .notModified:
-            Issue.record("Expected .updated for 200 response")
-        case let .updated(data: data, etag: _):
-            #expect(data.items.count == 1)
-            #expect(data.items.first?.name == "wget")
-            #expect(data.decodeFailures.count == 1)
-            #expect(data.decodeFailures.first?.index == 1)
-        }
-    }
 }
